Updated Marathon

+++ marathon/app/Controllers/Admin.php
@@ -2,6 +2,7 @@
 
 namespace App\Controllers;
 
+use App\Models\Member;
 use App\Models\Race;
 
 class Admin extends BaseController
@@ -14,6 +15,7 @@ class Admin extends BaseController
     }
     public function manage_marathon(): string
     {
+
         $Race = new Race();
         $data = array('manage_marathon'=>'true');
         $data['races'] = $Race->get_races();
@@ -46,23 +48,64 @@ class Admin extends BaseController
     public function delete_race($id)
     {
         $Race = new Race();
-        $Race->delete_race($id);
+        $this->session = service('session');
+        $this->session->start();
+        $memberKey = $this->session->get("memberKey");
+        $Member = new Member();
+        if( $Member->has_access($id, $memberKey))
+        {
+            $Race->delete_race($id);
+        }else{
+            echo "Error Access Denied";
+            exit();
+        }
+
         header('Refresh:0; url=/marathon/public/marathon');
         exit();
     }
     //loading view with the data that needs updating
     public function update_race($id)
     {
-        $Race = new Race();
         $data = array('manage_marathon'=>'true');
-        $data['race'] = $Race->get_race($id);
+        $this->session = service('session');
+        $this->session->start();
+        $memberKey = $this->session->get("memberKey");
+        $Member = new Member();
+        if( $Member->has_access($id, $memberKey))
+        {
+            $Race = new Race();
+            $data['race'] = $Race->get_race($id);
+        }else{
+            echo "Error Access Denied";
+            exit();
+        }
+
+
         return view('update_page', $data);
     }
     public function edit_race()
     {
         $Race = new Race();
-        $Race->update_race($this->request->getPost('race_name'), $this->request->getPost('race_location'), $this->request->getPost('race_description'), $this->request->getPost('race_date'), $this->request->getPost('txtID'));
+        $this->session = service('session');
+        $this->session->start();
+        $memberKey = $this->session->get("memberKey");
+        $Member = new Member();
+        if( $Member->has_access($this->request->getPost('txtID'), $memberKey))
+        {
+            $Race->update_race($this->request->getPost('race_name'), $this->request->getPost('race_location'), $this->request->getPost('race_description'), $this->request->getPost('race_date'), $this->request->getPost('txtID'));
+
+        }else{
+            echo "Error Access Denied";
+            exit();
+        }
         header('Refresh:0; url=/marathon/public/marathon');
+        exit();
+    }
+    public function logout()
+    {
+        $this->session = service('session');
+        $this->session->destroy();
+        header('Refresh:0; url=/marathon/public/');
         exit();
     }
 }
